apps/processor/requirements.txt
pdfplumber>=0.10.0
python-docx>=1.1.0
openpyxl>=3.1.0
pandas>=2.2.0
boto3>=1.34.0
Pillow>=10.2.0
anthropic>=0.52.0
reportlab>=4.1.0
stegano>=0.11.0
pytest>=8.0.0
